read commit msg before

+++ randomWalk/Cuda/mainkernels.h
@@ -6,7 +6,7 @@
 
 #define NSAMPLE 200
 
-void randomWalkCudaWrapper(int dimBlck,int dimThread,QuadTreeManager* quadTree, int RECT_ID,unsigned int *output,unsigned int randomSeed);
+void randomWalkCudaWrapper(int dimBlck,int dimThread,QuadTreeManager* quadTree, int RECT_ID,unsigned int *output,unsigned long long randomSeed);
 QuadTreeManager* createQuadTree(const std::vector<d_Rect>& layer,d_Rect const& spaceSize,bool doCheck);
 QuadTreeManager* randomWalkCudaInit(char* path);
 bool initCuda(int argc, char **argv);
